Cleanup export make_links function
Use ln link notation of
- link name for the path where the link resides
- link target for the path that the link points to
in hope to make it a bit less confusing to read.

Changed the exists() method previously used to check whether
the symlink existed, because when it does, the exists() method will
try to traverse the symlink to its target, to actually check whether
the symlink points to an existing target.