Try guessing the collective in the upload command
If directories are traversed with the `upload` command and `-i` is
specfified to use the integration endpoint, the collective is now
deduced from the file being uploaded. This is the same as done in the
`watch` command (and also applies now to `cleanup` and `file-exists`).
If single files are uploaded, a collective must be given with the
arguments.

Fixes: #125